requirements.txt
pyserial
pyrtcm==1.0.14
pyubx2==1.2.35
pygnssutils==1.0.16
